% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{get_gaa}
\alias{get_gaa}
\title{Download GAA documents}
\usage{
get_gaa(year = "all")
}
\arguments{
\item{year}{A list (or coercible to one) specifying the fiscal year(s) to download.
The default value is \code{"all"} the years available.
If more than one \code{type} is set, a single \code{year} value or vector will be recycled.}
}
\value{
\code{NULL}
}
\description{
A wrapper for \code{get_docs(type = "gaa")}; this function downloads the
General Appropriations Act (GAA) Excel files (.xlsx) from the
Department of Budget and Management (DBM) website
}
\examples{
\dontrun{
  get_gaa() # Downloads all GAA files from 2020 onwards
}
}
