% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{get_docs}
\alias{get_docs}
\title{Bulk download budget documents}
\usage{
get_docs(type = "all", year = "all")
}
\arguments{
\item{type}{A character string or vector specifying the type of budget document(s) to download:
\itemize{
\item \code{"gaa"}; General Appropriations Act
\item \code{"nep"}; National Expenditure Program
\item \code{"all"} (the default); equivalent to \code{c("gaa", "nep")}
}}

\item{year}{A list (or coercible to one) specifying the fiscal year(s) to download.
The default value is \code{"all"} the years available.
If more than one \code{type} is set, a single \code{year} value or vector will be recycled.}
}
\value{
Invisibly returns \code{NULL} (called for side effects).
}
\description{
Download the following raw budget documents in Excel format (.xlsx)
from the Department of Budget and Management (DBM) website:
\itemize{
\item National Expenditure Program (NEP)
\item General Appropriations Act (GAA)
}
}
\examples{
\dontrun{
  get_docs()  # downloads all GAA and NEP Excel files from all the years available

  # equivalent calls----
  get_docs("all", 2020:2024) # the year value will be recycled for each type
  get_docs(c("gaa", "nep"), list(2020:2024, 2020:2024))

}
}
\seealso{
\code{\link[=get_gaa]{get_gaa()}} to download GAA documents
\code{\link[=get_nep]{get_nep()}} to download NEP documents
}
