% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{get_nep}
\alias{get_nep}
\title{Download NEP documents}
\usage{
get_nep(year = NULL)
}
\arguments{
\item{year}{A list (or coercible to one) specifying the fiscal year(s) to download.
The default value is \code{NULL} or all the years supported for download. (currently 2020 up to the most recent fiscal year)
If more than one \code{type} is set, a single \code{year} value or vector will be recycled.}
}
\value{
\code{NULL}
}
\description{
#' A wrapper for \code{get_docs(type = "nep")}; this function downloads the
General Appropriations Act (GAA) Excel files (.xlsx) from the
Department of Budget and Management (DBM) website
}
\examples{
\dontrun{
get_nep() # Downloads all NEP files from 2020 onwards
}
}
